- Eliminado código obsoleto al asignar un nuevo código a las facturas de compra. - Ya se pueden seleccionar más sistemas de envío de emails: mail, sendmail y smtp. - Añadidas más comprobaciones al informa de errores para solucionar errores en la tabla de clientes. - Al crear nuevos clientes o proveedores ya se puede especificar si es una persona física (no una empresa). - Solucionado bug al mostrar la suma de los netos en el asistente de agrupar albaranes de venta. - El algoritmo de comprobación de duplicados de facturas comprueba ahora también numero2 y numproveedor. - Solucionado bug al eliminar una regularización de IVA, ahora se elimina el asiento al final, evita errores. - Solucionado bug al eliminar artículos con imagen. No se estaba eliminando la imagen.

+++ model/core/regularizacion_iva.php
@@ -205,18 +205,25 @@ class regularizacion_iva extends \fs_model
    
    public function delete()
    {
-      /// si hay un asiento asociado lo eliminamos
-      if( isset($this->idasiento) )
+      if( $this->db->exec("DELETE FROM ".$this->table_name." WHERE idregiva = ".$this->var2str($this->idregiva).";") )
       {
-         $asiento = new \asiento();
-         $as0 = $asiento->get($this->idasiento);
-         if($as0)
+         /// si hay un asiento asociado lo eliminamos
+         if( isset($this->idasiento) )
          {
-            $as0->delete();
+            $asiento = new \asiento();
+            $as0 = $asiento->get($this->idasiento);
+            if($as0)
+            {
+               $as0->delete();
+            }
          }
+         
+         return TRUE;
       }
-      
-      return $this->db->exec("DELETE FROM ".$this->table_name." WHERE idregiva = ".$this->var2str($this->idregiva).";");
+      else
+      {
+         return FALSE;
+      }
    }
    
    /**
